+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-
+- required packages and client and server endpoints
 
 ### Changed
 
+
+[Unreleased]: https://github.com/bohrsty/agenda-display/compare/81c0b52...HEAD